# swagger.json
{
    "swagger":"2.0",
    "info": {
        "version" : "1.0.0",
        "title" : "Ecommerce App",
        "description" : "My Ecommcerce Project API",
        "license" : {
            "name" : "MIT",
            "url":"https://opensource.org/licenses/MIT"
        }
    },
    "host" : "localhost:8000",
    "basePath" : "/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces" : ["application/json"],
    "paths" : {
...
        "/api/user/{id}" :{
            "get" :{
                "tags" : ["User"],
                "summary" : "Get User have ID",
                "parameters":[
                    {
                        "in":"path",
                        "name":"id",
                        "description":"User ID",
                        "required": true,
                        "schema":{
                            "type":"string"
                        }
                    }],
                "responses":{
                    "200" : {
                        "message": "OK",
                        "description" : "OK",
                        "schema":{
                            "$ref" :  "#/definitions/UserResponse"
                        }
                    }
                }
            }
        
        },
        "/api/user/add_user":{
            "post":{
                "tags" : ["User"],
                "summary":"Add One User",
                "parameters":[
                    {
                        "name": "body",
                        "in" : "body",
                        "description" : "Your Username",
                        "required": true,
                        "schema" :{
                            "properties":{
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "url": {
                                    "type": "string"
                                },
                                "userType": {
                                    "type": "string"
                                }
                            }
                        }
                    }

                ],
                "responses":{
                    "200":{
                            "message": "OK",
                            "description" : "OK",
                            "schema":{
                                "$ref" :  "#/definitions/UserResponse"
                            }
                    }
                }
            }
        },
        "/api/user/login":{
            "post":{
                "tags" : ["User"],
                "summary":"Login to System",
                "parameters":[
                    {
                        "name": "body",
                        "in" : "body",
                        "description" : "Your Account",
                        "required": true,
                        "schema" :{
                            "properties":{
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }

                ],
                "responses":{
                    "200":{
                            "message": "OK",
                            "description" : "OK",
                            "schema":{
                                "$ref" :  "#/definitions/UserResponse"
                            }
                    }
                }
            }
        },
        "/api/user/find_google_account":{
            "post":{
                "tags" : ["User"],
                "summary":"Find Google User in System",
                "parameters":[
                    {
                        "name": "body",
                        "in" : "body",
                        "description" : "Your Account",
                        "required": true,
                        "schema" :{
                            "properties":{
                                "username": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                }
                            }
                        }
                    }

                ],
                "responses":{
                    "200":{
                            "message": "OK",
                            "description" : "OK",
                            "schema":{
                                "$ref" :  "#/definitions/UserResponse"
                            }
                    }
                }
            }
        },
        "/api/user/update/{userID}":{
            "put":{
                "tags" : ["User"],
                "summary":"Update User in System",
                "parameters":[
                    {
                        "in":"path",
                        "name":"userID",
                        "description":"User ID",
                        "required": true,
                        "schema":{
                            "type":"string"
                        }
                    },
                    {
                        "name": "body",
                        "in" : "body",
                        "description" : "Your Account",
                        "required": true,
                        "schema" :{
                            "properties":{
                                "email": {
                                    "type": "string"
                                },
                                "address":{
                                    "type":"string"
                                },
                                "phone":{
                                    "type":"string"
                                },
                                "url":{
                                    "type":"string"
                                }
                            }
                        }
                    }

                ],
                "responses":{
                    "200":{
                            "message": "OK",
                            "description" : "OK",
                            "schema":{
                                "$ref" :  "#/definitions/UserResponse"
                            }
                    }
                }
            }
        },
        "/api/user/update_cart/{id}":{
            "put":{
                "tags" : ["User"],
                "summary":"Update Cart For User In System",
                "parameters":[
                    {
                        "in":"path",
                        "name":"id",
                        "description":"User ID",
                        "required": true,
                        "schema":{
                            "type":"string"
                        }
                    },
                    {
                        "name": "body",
                        "in" : "body",
                        "description" : "Your Account",
                        "required": true,
                        "schema" :{
                            "properties":{
                                "cartID": {
                                    "type": "string"
                                }
                            }
                        }
                    }

                ],
                "responses":{
                    "200":{
                            "message": "OK",
                            "description" : "OK",
                            "schema":{
                                "$ref" :  "#/definitions/UserResponse"
                            }
                    }
                }
            }
        },
...
    },
    "definitions" : {
        "user" : {
            "required" : ["userID","username","password"],
            "properties" : {
                "_id" : {
                    "type" : "string"
                },
                "userID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "login_type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "cartID": {
                    "type": "string"
                },
                "user_type":{
                    "type":"string"
                }
            }
        },
...
        "UserResponse" : {
            "properties" : {
                "message" : {
                    "type" : "String"
                },
                "data" : {
                    "type" : "array",
                    "items" : {
                        "$ref" :"#/definitions/user"
                    }
                }
            }
        }
    }
}